Split classes from core/renderers/measurables/inputs.js
